coolgoat/requirements.txt
asgiref==3.8.1
Authlib==1.3.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
Django==5.1
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
ecdsa==0.19.0
idna==3.10
mock==5.1.0
psycopg2==2.9.9
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
python-jose==3.3.0
requests==2.32.3
rsa==4.9
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
paho-mqtt==1.6.1
